refactor: move deserialization tests inline, remove dead code
- Move 10 deserialization tests from integration test file into
  gemini.rs #[cfg(test)] module — eliminates duplicated struct
  definitions (5 types + impl block + extract_text helper)
- Keep only 2 API integration tests in the separate test file
- Remove redundant body.error check after 2xx (dead code since
  non-2xx is now handled before JSON parsing)
- Strengthen thinking-response test to assert exact extracted text
- Extract require_api_key() helper for integration tests
- Remove unused serde/serde_json dev-dependencies
