+++ src-tauri/llm/Cargo.toml
@@ -12,5 +12,3 @@ anyhow = "1.0"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
