+++ src-tauri/llm/src/gemini.rs
@@ -159,10 +159,6 @@ pub async fn generate_branch_name_gemini(
             .await
             .context("Failed to parse Gemini response")?;
 
-        if let Some(error) = body.error {
-            anyhow::bail!("Gemini API error ({}): {}", status, error.message());
-        }
-
         let raw_text = body
             .candidates
             .and_then(|c| c.into_iter().next())
@@ -191,6 +187,20 @@ pub async fn generate_branch_name_gemini(
 mod tests {
     use super::*;
 
+    fn extract_text(body: &GeminiResponse) -> Option<String> {
+        body.candidates
+            .as_ref()
+            .and_then(|c| c.first())
+            .and_then(|c| c.content.as_ref())
+            .and_then(|content| {
+                content
+                    .parts
+                    .iter()
+                    .find(|p| p.thought != Some(true))
+                    .and_then(|p| p.text.clone())
+            })
+    }
+
     #[test]
     fn test_system_prompt_not_empty() {
         assert!(!SYSTEM_PROMPT.is_empty());
@@ -217,5 +227,167 @@ mod tests {
         let json = serde_json::to_string(&request).unwrap();
         assert!(json.contains("system_instruction"));
         assert!(json.contains("generation_config"));
+    }
+
+    #[test]
+    fn standard_gemini_2_response_parses() {
+        let json = r#"{
+            "candidates": [{
+                "content": {
+                    "parts": [{"text": "feat/add-oauth-auth"}],
+                    "role": "model"
+                },
+                "finishReason": "STOP",
+                "safetyRatings": [{"category": "HARM_CATEGORY_SEXUALLY_EXPLICIT", "probability": "NEGLIGIBLE"}]
+            }],
+            "usageMetadata": {"promptTokenCount": 50, "candidatesTokenCount": 10, "totalTokenCount": 60}
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("standard response should parse");
+        assert_eq!(
+            extract_text(&response),
+            Some("feat/add-oauth-auth".to_string())
+        );
+    }
+
+    #[test]
+    fn error_response_with_object_parses() {
+        let json = r#"{
+            "error": {
+                "code": 429,
+                "message": "Resource has been exhausted (e.g. check quota).",
+                "status": "RESOURCE_EXHAUSTED"
+            }
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("error response should parse");
+        assert!(response.error.is_some());
+        assert!(response.error.unwrap().message().contains("exhausted"));
+    }
+
+    #[test]
+    fn model_not_found_error_parses() {
+        let json = r#"{
+            "error": {
+                "code": 404,
+                "message": "models/nonexistent-model is not found for API version v1beta.",
+                "status": "NOT_FOUND"
+            }
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("404 error should parse");
+        assert!(response.error.unwrap().message().contains("not found"));
+    }
+
+    /// Bug #446: Gemini 3 returns thinking parts alongside text parts.
+    /// extract_text must skip thought parts and return the real text.
+    #[test]
+    fn gemini_3_thinking_response_skips_thought_parts() {
+        let json = r#"{
+            "candidates": [{
+                "content": {
+                    "parts": [
+                        {"thought": true, "text": "Let me think about a good branch name..."},
+                        {"text": "feat/add-oauth-auth"}
+                    ],
+                    "role": "model"
+                },
+                "finishReason": "STOP"
+            }],
+            "modelVersion": "gemini-3-flash-preview"
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("thinking response should parse");
+        assert_eq!(
+            extract_text(&response),
+            Some("feat/add-oauth-auth".to_string())
+        );
+    }
+
+    /// Bug #446: thoughtSignature fields should be ignored by serde.
+    #[test]
+    fn gemini_3_response_with_thought_signatures_parses() {
+        let json = r#"{
+            "candidates": [{
+                "content": {
+                    "parts": [
+                        {"thought": true, "text": "analyzing..."},
+                        {"text": "feat/add-oauth-auth", "thoughtSignature": "c2lnbmF0dXJlX2RhdGE="}
+                    ],
+                    "role": "model"
+                },
+                "finishReason": "STOP"
+            }],
+            "modelVersion": "gemini-3-flash-preview"
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("response with thoughtSignature should parse");
+        assert_eq!(
+            extract_text(&response),
+            Some("feat/add-oauth-auth".to_string())
+        );
+    }
+
+    /// Bug #446: Safety-blocked candidates have no `content` field.
+    #[test]
+    fn candidate_without_content_safety_blocked_parses() {
+        let json = r#"{
+            "candidates": [{
+                "finishReason": "SAFETY",
+                "safetyRatings": [
+                    {"category": "HARM_CATEGORY_SEXUALLY_EXPLICIT", "probability": "HIGH"}
+                ]
+            }]
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("safety-blocked candidate should parse");
+        assert_eq!(extract_text(&response), None);
+    }
+
+    /// Bug #446: MAX_TOKENS candidates may have no content (all tokens used by thinking).
+    #[test]
+    fn candidate_without_content_max_tokens_parses() {
+        let json = r#"{
+            "candidates": [{
+                "finishReason": "MAX_TOKENS"
+            }],
+            "modelVersion": "gemini-3-flash-preview"
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("MAX_TOKENS candidate should parse");
+        assert_eq!(extract_text(&response), None);
+    }
+
+    /// Bug #446: Error as plain string instead of object.
+    #[test]
+    fn error_as_string_parses() {
+        let json = r#"{
+            "error": "Model gemini-3-flash-preview is not available."
+        }"#;
+
+        let response: GeminiResponse =
+            serde_json::from_str(json).expect("string error should parse");
+        assert_eq!(
+            response.error.unwrap().message(),
+            "Model gemini-3-flash-preview is not available."
+        );
+    }
+
+    #[test]
+    fn non_json_response_fails_deserialization() {
+        let html = r#"<html><body><h1>502 Bad Gateway</h1></body></html>"#;
+        assert!(serde_json::from_str::<GeminiResponse>(html).is_err());
+    }
+
+    #[test]
+    fn empty_response_fails_deserialization() {
+        assert!(serde_json::from_str::<GeminiResponse>("").is_err());
     }
 }
